Actualizando la calculadora con los requisitos de el ejercicio

+++ Calculadora/src/Multiplicacion.java
@@ -1,11 +1,11 @@
 public class Multiplicacion extends Operaciones {
-    Multiplicacion(int numero1, int numero2){
+    Multiplicacion(double numero1, double numero2){
         super(numero1, numero2);
     }
 
     @Override
     public void operar(){
-        int resultado = num1 * num2;
+        double resultado = num1 * num2;
         System.out.println("La multiplicacion entre el numero " + num1 + " y el numero " + num2 + " es: " + resultado);
     }
 }
