+++ Calculadora/src/Operaciones.java
@@ -1,8 +1,8 @@
 public abstract class Operaciones {
-    public int num1;
-    public int num2;
+    public double num1;
+    public double num2;
 
-    Operaciones(int num1, int num2){
+    Operaciones(double num1, double num2){
         this.num1 = num1;
         this.num2 = num2;
     }
