+++ Calculadora/src/Resta.java
@@ -1,12 +1,12 @@
 public class Resta extends Operaciones {
 
-    Resta(int numero1, int numero2){
+    Resta(double numero1, double numero2){
         super(numero1, numero2);
     }
 
     @Override
     public void operar() {
-        int resultado = num1 - num2;
+        double resultado = num1 - num2;
         System.out.println("La resta de el numero " + num1 + " y el numero " + num2 + " es: " + resultado);
     }
 }
